+++ src/main/java/com/mallang/mallang_backend/domain/video/video/service/cache/VideoCacheEntry.java
@@ -1,22 +1,20 @@
-package com.mallang.mallang_backend.domain.video.video.service.cache;
-
-import java.util.List;
-
-import com.mallang.mallang_backend.domain.video.video.dto.VideoResponse;
-
-import lombok.Getter;
-
-@Getter
-public class VideoCacheEntry {
-	private List<VideoResponse> videos;
-	private long fetchedMaxResults;
-
-	// 기본 생성자(직렬화용)
-	public VideoCacheEntry() {}
-
-	public VideoCacheEntry(List<VideoResponse> videos, long fetchedMaxResults) {
-		this.videos = videos;
-		this.fetchedMaxResults = fetchedMaxResults;
-	}
-
-}
+// package com.mallang.mallang_backend.domain.video.video.service.cache;
+//
+// import java.util.List;
+//
+// import com.mallang.mallang_backend.domain.video.video.dto.VideoResponse;
+//
+// import lombok.AllArgsConstructor;
+// import lombok.Getter;
+// import lombok.NoArgsConstructor;
+//
+// /**
+//  * 캐시된 비디오 목록과 가져온 최대 결과 수를 저장하는 클래스
+//  */
+// @Getter
+// @NoArgsConstructor
+// @AllArgsConstructor
+// public class VideoCacheEntry {
+// 	private List<VideoResponse> videos;
+// 	private long fetchedMaxResults;
+// }
